Allows to compile p384 with arch others than arm64 and amd64.
