+++ ecc/p384/arith_amd64.s
@@ -1,4 +1,4 @@
-// +build amd64
+// +build amd64,!noasm
 
 #include "textflag.h"
